ocr-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
paddleocr==3.0.0
paddlepaddle==3.0.0
python-multipart==0.0.20
Pillow==11.2.1
PyMuPDF==1.25.5

# dev / test
pytest==8.3.5
httpx==0.28.1
